Renamed "ButtAPI" to "butt_api". Uppercase files names are buggy in dota.
